! can: make Content-Length a long value, fixes #443
API changes:
  * `Content-Length`.length is now a Long
  * `max-content-length` setting is now also a Long
  * `incoming-auto-chunking-threshold-size` is now also a Long

This change makes only sense in combination with (auto-)chunking,
because we are not able to support byte arrays (HttpEntities) bigger
than Int.MaxValue.

+++ spray-http/src/main/scala/spray/http/HttpHeader.scala
@@ -217,7 +217,7 @@ object HttpHeaders {
   }
 
   object `Content-Length` extends ModeledCompanion
-  case class `Content-Length`(length: Int)(implicit ev: ProtectedHeaderCreation.Enabled) extends ModeledHeader {
+  case class `Content-Length`(length: Long)(implicit ev: ProtectedHeaderCreation.Enabled) extends ModeledHeader {
     def renderValue[R <: Rendering](r: R): r.type = r ~~ length
     protected def companion = `Content-Length`
   }
